Extract provider route cooldown helpers
